+++ app/src/main/java/org/oxycblt/auxio/music/Music.kt
@@ -23,8 +23,6 @@ import android.content.Context
 import android.net.Uri
 import android.provider.MediaStore
 import org.oxycblt.auxio.R
-import org.oxycblt.auxio.music.backend.id3v2GenreName
-import org.oxycblt.auxio.music.backend.withoutArticle
 import org.oxycblt.auxio.ui.Item
 import org.oxycblt.auxio.util.unlikelyToBeNull